9.4.0 (#309)
* feat: specialize-into-bytes ff
* feat: add scan_buffered and scan_cluster_buffered
* feat: auto continue scanning on page drop

---------

+++ src/commands/impls/server.rs
@@ -150,7 +150,7 @@ pub async fn ping<C: ClientLike>(client: &C) -> Result<RedisValue, RedisError> {
 }
 
 pub async fn select<C: ClientLike>(client: &C, db: u8) -> Result<RedisValue, RedisError> {
-  let frame = utils::request_response(client, || Ok((RedisCommandKind::Select, vec![db.into()]))).await?;
+  let frame = utils::request_response(client, || Ok((RedisCommandKind::Select, vec![(db as i64).into()]))).await?;
   protocol_utils::frame_to_results(frame)
 }
